frontend: Strip URLs before validating
For some weird reason the Django validator does not seem to
like trainling white space.  We don't care so much about that
but rather about making it easy for the user to submit lists.
And it's quite possible that the user has URLs with trailing whitespace.
The browser strips such URLs and so should we.